Changes and comments
