+++ server/jsonutils/path.go
@@ -117,74 +117,3 @@ func FormatPrefixSuffix(key string) string {
 	}
 	return key
 }
-
-type JSONPathes struct {
-	pathes map[string]*JSONPath
-}
-
-func NewJSONPathes(pathes []string) *JSONPathes {
-	container := make(map[string]*JSONPath)
-
-	for _, path := range pathes {
-		container[path] = NewJSONPath(path)
-	}
-
-	return &JSONPathes{
-		pathes: container,
-	}
-}
-
-func (jpa *JSONPathes) String() string {
-	result := ""
-
-	for key := range jpa.pathes {
-		if result != "" {
-			result += ", "
-		}
-		result += key
-	}
-
-	return "[" + result + "]"
-}
-
-func (jpa *JSONPathes) Get(object map[string]interface{}) (map[string]interface{}, bool) {
-	result := false
-	array := make(map[string]interface{})
-
-	for key, path := range jpa.pathes {
-		value, answer := path.Get(object)
-		array[key] = value
-		result = result || answer
-	}
-
-	return array, result
-}
-
-func (jpa *JSONPathes) Set(object map[string]interface{}, values map[string]interface{}) error {
-	for key, path := range jpa.pathes {
-		value := values[key]
-		if value != nil {
-			err := path.Set(object, value)
-			if err != nil {
-				return err
-			}
-		}
-	}
-
-	return nil
-}
-
-func (jpa *JSONPathes) IsFullFilled(object map[string]interface{}) bool {
-	result := true
-
-	for _, path := range jpa.pathes {
-		value, answer := path.Get(object)
-		result = result && answer && value != nil
-
-		if !result {
-			return result
-		}
-	}
-
-	return result
-}
